Track deletions more carefully.
Mercurial is very picky about what exists and what doesn't! I noticed
this because on nodachi, which is (at the moment) lacking a number of
meta.*.txt files, deleting a page caused a "R"emoved file to be listed
in "hg st" output at the command-line, and a manual commit was
needed. The problem was Mercurial's pickiness. The solution is to only
try to "hg rm" and "hg ci" the files that were *actually* removed
during the given transaction - files that a failed removal attempt was
made for (e.g. because of nonexistence or permissions problems) aren't
passed on through to Mercurial.

darcs-hash:20080523181131-e82e2-16421a093737e91d0f33862fdc79ac604c893103.gz